2014-02-24 Roman - Transport: Extended parameters prm1 and prm2 type of output transport is changet to int64_t. - FIX:TConfig,DB:{FireBird,PostgreSQL,SQLite}: Moving to int64_t for integer is fixed. - FIX:DB.{MySQL,PostgreSQL}: Context help fix. - DB.SQLite: Requests to DB the resources allocation is changed to use pthread_mutex_t. Mostly code clean up. - TArchive,Archive.DBArch: To TVArchEl::setValsProc() added result for allow archive's buffer's part write control.     Used for rewrite to network DB by DBArch. - Archive.DBArch: Allow start Values and Messages archivators for DB error, predict restore the connection. - DAQ.AMRDevs: Kontar implementing further progress. - FIX:DAQ.OPC_UA: OpcUa_Server_ServerArray variable process append. epEnAt() result process append for some crash prevent. - FIX:UI.WebVision: Posible crash into primitive ElFig fix.
git-svn-id: svn://oscada.org/trunk/OpenSCADA@2055 7bc60720-51fb-0310-bb35-cf765b5d350c

+++ src/tconfig.cpp
@@ -237,7 +237,7 @@ TCfg::TCfg( TFld &fld, TConfig &owner ) : mView(true), mKeyUse(false), mNoTransl
     switch(mFld->type())
     {
 	case TFld::String:	setType(TVariant::String, true);  TVariant::setS(mFld->def());	break;
-	case TFld::Integer:	setType(TVariant::Integer, true); TVariant::setI(atoi(mFld->def().c_str()));	break;
+	case TFld::Integer:	setType(TVariant::Integer, true); TVariant::setI(atoll(mFld->def().c_str()));	break;
 	case TFld::Real:	setType(TVariant::Real, true);	  TVariant::setR(atof(mFld->def().c_str()));	break;
 	case TFld::Boolean:	setType(TVariant::Boolean, true); TVariant::setB((bool)atoi(mFld->def().c_str()));break;
 	default: break;
@@ -302,7 +302,7 @@ void TCfg::setS( const string &val )
 {
     switch(type())
     {
-	case TVariant::Integer:	setI(atoi(val.c_str()));	break;
+	case TVariant::Integer:	setI(atoll(val.c_str()));	break;
 	case TVariant::Real:	setR(atof(val.c_str()));	break;
 	case TVariant::Boolean:	setB((bool)atoi(val.c_str()));	break;
 	case TVariant::String:
